refactor reporting units endpoint
