+++ ras_backstage/__init__.py
@@ -25,6 +25,7 @@ collection_exercise_api = Namespace('collection-exercise',
 collection_instrument_api = Namespace('collection-instrument',
                                       path='/backstage-api/v1/collection-instrument')
 party_api = Namespace('party', path='/backstage-api/v1/party')
+reporting_unit_api = Namespace('reporting-unit', path='/backstage-api/v1/reporting-unit')
 sample_api = Namespace('sample', path='/backstage-api/v1/sample')
 secure_messaging_api = Namespace('secure-messaging',
                                  path='/backstage-api/v1/secure-message')
@@ -35,6 +36,7 @@ survey_api = Namespace('survey', path='/backstage-api/v1/survey')
 api.add_namespace(collection_exercise_api)
 api.add_namespace(collection_instrument_api)
 api.add_namespace(party_api)
+api.add_namespace(reporting_unit_api)
 api.add_namespace(sample_api)
 api.add_namespace(secure_messaging_api)
 api.add_namespace(sign_in_api)
@@ -47,7 +49,7 @@ from ras_backstage.resources.collection_exercise.single_collection_exercise impo
 from ras_backstage.resources.collection_instrument.collection_instrument import CollectionInstrument  # NOQA # pylint: disable=wrong-import-position
 from ras_backstage.resources.info import Info  # NOQA # pylint: disable=wrong-import-position
 from ras_backstage.resources.party.get_party_details import PartyDetails  # NOQA # pylint: disable=wrong-import-position
-from ras_backstage.resources.party.search_businesses import SearchBusinesses  # NOQA # pylint: disable=wrong-import-position
+from ras_backstage.resources.reporting_units.search_reporting_units import SearchReportingUnits  # NOQA # pylint: disable=wrong-import-position
 from ras_backstage.resources.sample.sample import Sample  # NOQA # pylint: disable=wrong-import-position
 from ras_backstage.resources.secure_messaging.get_message_list import GetMessagesList  # NOQA # pylint: disable=wrong-import-position
 from ras_backstage.resources.secure_messaging.get_message import GetMessage  # NOQA # pylint: disable=wrong-import-position
