+++ ras_backstage/resources/reporting_units/search_reporting_units.py
@@ -4,7 +4,7 @@ from flask import jsonify, make_response, request
 from flask_restplus import Resource, reqparse
 from structlog import wrap_logger
 
-from ras_backstage import party_api
+from ras_backstage import reporting_unit_api
 from ras_backstage.controllers import party_controller
 
 
@@ -14,17 +14,17 @@ parser = reqparse.RequestParser()
 parser.add_argument('query', location='args', required=True)
 
 
-@party_api.route('/search-businesses')
-class SearchBusinesses(Resource):
+@reporting_unit_api.route('/search')
+class SearchReportingUnits(Resource):
 
     @staticmethod
-    @party_api.expect(parser)
+    @reporting_unit_api.expect(parser)
     def get():
         query = request.args.get('query')
 
-        logger.info('Retrieving businesses by search query', query=query)
+        logger.info('Retrieving reporting units by search query', query=query)
 
         businesses = party_controller.get_businesses_by_search(query)
 
-        logger.info('Successfully retrieved businesses by search query')
+        logger.info('Successfully retrieved reporting units by search query')
         return make_response(jsonify(businesses), 200)
